(I will write my commit messages in english from now on)
I have connected the ToDoService to the Data Access Layer (DAL).

I have implemented the GetToDo method that accept a string with a name as a parameter.

I have put some test code in the console application to show that the method works, it will retrieve the items with the name "Hamid".

This commit does not give the console application a "done" status for the GetToDo functionality, it should let the user choose the function from a menu and let the user enter the name of the list to retrieve.