Forgot to write in earlier commit message: I added UNIT testing to the ToDo class, it is in the ToDoBaseTests project. It contains some testing code to validate the IComparable interface implementation.
Added code that stops AddToDo from inserting duplicate ToDo items into the database. The comparision is done on Name, Description, DeadLine and Estimation time. ID, CreatedDate, Finnished is ignored by design.

Added code in AddToDo to check that the new ToDo item made it to the database.

I took the liberty of deciding that the "add points to todo list" user story is taken care of, as adding a point to the list is the same method as creating a new list by adding a point.

TODO in the future: Decide how to handle feedback to the caller for failed database insertions in a RESTful way.

Refactor how the "did the todo list make it into the database" check is done, instead of reading all todos from the database, there is an error code in the DAL object property "ErrorMessage" that should be checked. However, the code need to be looked at, as the ErrorMessage might be an old one. Remember, we are using the same instance over and over again...